Create Challenge Stage component
