+++ frontend/src/components/challenges/ChallengeBattle.jsx
@@ -1,0 +1,6 @@
+function ChallengeBattle() {
+  return (
+    <div>ChallengeBattle</div>
+  )
+}
+export default ChallengeBattle
